server/requirements.txt
firebase-admin
flask
python-dotenv
flask-jwt-extended